maitri/ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "espeak==1.48.4",
    "fastapi==0.104.1",
    "gtts==2.4.0",
    "librosa==0.10.1",
    "loguru==0.7.2",
    "mediapipe==0.10.7",
    "numpy==1.24.3",
    "onnxruntime==1.16.3",
    "opencv-python==4.8.1.78",
    "pillow==10.0.1",
    "pyaudio==0.2.11",
    "pydantic==2.4.2",
    "pydub==0.25.1",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "scikit-learn==1.3.0",
    "scipy==1.11.3",
    "speechrecognition==3.10.0",
    "sqlalchemy==2.0.23",
    "torch==2.1.0",
    "torchaudio==2.1.0",
    "torchvision==0.16.0",
    "transformers==4.35.2",
    "uvicorn==0.24.0",
]
